Added Dockerfile for RMS Docker image
Creates Restcomm Media Server docker image that runs without any of the bash startup script helpers, to reduce complexity and tech debt.
This means that it relies entirely on some form of externalized configuration, such as a docker volume for each of the config files (media-server.yml, log4j2.yml, media-plugin-vad-noise-threshold.yml, etc.)

With this approach, we are aiming to extract any kind of business logic that has to do with the *configuration* of the application outside of the *binary* (the docker image). In addition, this has the side benefit that configuration changes don't require building a new binary.