Added docker-compose for RC - RMS
An adapted version of the Restcomm-Connect `docker-compose.yml` file, which also deploys the separate RMS docker container.

The docker-compose is aimed for deployment in a Vagrant-powered Ubuntu VM, which has a specific static IP. Using Vagrant also allows us to avoid dealing with "Docker for Mac" networking issues related to "host networking".